Extracted AssetSelector from FloormapDisplay component
