+++ src/pages/FloormapDetail.jsx
@@ -1,13 +1,19 @@
-import React, { useEffect, useState } from "react";
+import React, { useState, useEffect, useCallback } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import { FloorMapService } from "../services/floormapService.js";
 import AssetSimulationService from "../services/assetSimulationService.js";
 import FloormapDisplay from "../components/FloormapDisplay/FloormapDisplay.jsx";
+import AssetSelector from "../components/AssetSelector/AssetSelector.jsx";
 
 function FloormapDetail() {
   const { floormapId } = useParams();
   const navigate = useNavigate();
   const [assets, setAssets] = useState([]);
+  const [activeAsset, setActiveAssetState] = useState([]);
+
+  const setActiveAsset = useCallback((asset) => {
+    setActiveAssetState(asset);
+  }, []);
 
   useEffect(() => {
     FloorMapService.getFloorMapById(floormapId).then((floormap) => {
@@ -31,7 +37,13 @@ function FloormapDetail() {
 
   return (
     <div>
-      <FloormapDisplay floormapId={floormapId} assets={assets} />
+      <AssetSelector assets={assets} setActiveAsset={setActiveAsset} />
+      <FloormapDisplay
+        floormapId={floormapId}
+        assets={assets}
+        activeAsset={activeAsset}
+        setActiveAsset={setActiveAsset}
+      />
       <button onClick={handleEditZone} style={{ marginTop: "20px" }}>
         Edit Zones
       </button>
